requirements.txt
pyobjc>=8.0
requests
python-dotenv
langgraph
langchain-openai
langchain
sqlalchemy
typing
pydantic
langchain-community
langchain_experimental
neo4j
faiss-cpu
langchain_core
pydantic==2.7.4
pandas
numpy
opensearch-py